uppsrc: Fixed more Win64 warnings
git-svn-id: svn://ultimatepp.org/upp/trunk@5822 f0d560ea-af0d-0410-9eb7-867de7ffcac7

+++ uppsrc/Core/lheap.cpp
@@ -186,7 +186,7 @@ void *Heap::LAlloc(size_t& size) {
 		ptr = TryLAlloc(bini, size);
 		if(ptr) return ptr;
 	}
-	DLink *n = AddChunk(size);
+	DLink *n = AddChunk((int)size);
 	if(!n)
 		Panic("Out of memory!");
 	ptr = DivideBlock(n, (int)size, LBINS - 1);
